+++ CHANGELOG.md
@@ -3,6 +3,8 @@ GAMSPy CHANGELOG
 
 GAMSPy 1.7.0
 ------------
+- General
+  - Allow container serialization/deserialization.
 - Documentation
   - Add a section in FAQ about the compatibiltiy issues of the Python interpreter from the Microsoft Store.
 - Testing